consolidating code into app.js file and adjusting file structure
