--- src/components/PortfolioContainer.js
@@ -1,37 +1,0 @@
-import React, { useState } from 'react';
-import Navbar from './Navbar';
-import Home from './pages/Home';
-import About from './pages/About';
-import Resume from './pages/Resume';
-import Inquire from './pages/Inquire';
-import Work from './pages/Work';
-
-
-export default function PortfolioContainer() {
-  const [currentPage, setCurrentPage] = useState('Home');
-
-  const renderPage = () => {
-    if (currentPage === 'Home') {
-      return <Home />;
-    }
-    if (currentPage === 'About') {
-      return <About />;
-    }
-    if (currentPage === 'Resume') {
-      return <Resume />;
-    }
-    if (currentPage === 'Work') {
-      return <Work />;
-    }
-    return <Inquire />;
-  };
-
-  const handlePageChange = (page) => setCurrentPage(page);
-
-  return (
-    <div>
-      <Navbar currentPage={currentPage} handlePageChange={handlePageChange}/>
-      {renderPage()}
-    </div>
-  );
-}
